Fix --sort-memory

+++ src/AlignSettings.h
@@ -34,7 +34,7 @@ struct AlignSettings : MM2Settings
 
     bool Sort;
     int SortThreads;
-    int SortMemory;
+    int64_t SortMemory;
 
     /// Parses the provided CLI::Results and retrieves a defined set of options.
     AlignSettings(const PacBio::CLI::Results& options);
