+++ src/bam_sort.c
@@ -2289,7 +2289,7 @@ void sam_global_args_free(sam_global_args *ga)
 }
 
 int bam_sort(const char *inputName, const char *outputName, int numThreads, int merge_threads,
-             int memory, int *numFiles, int *numBlocks)
+             size_t memory, int *numFiles, int *numBlocks)
 {
     size_t max_mem = memory;
     int is_by_qname = 0;
